requirements.txt
numpy
opencv-python-headless
pillow
tesserocr
imutils
